requirements.txt
# Core dependencies
rdflib>=6.2.0
SPARQLWrapper>=2.0.0
pyyaml>=6.0

# Natural language processing
nltk>=3.8
spacy>=3.5.0

# Development dependencies
pytest>=7.2.0
pytest-cov>=4.0.0
black>=23.0.0
mypy>=1.0.0
flake8>=6.0.0

# Optional dependencies for pattern enrichment
wordnet>=0.1.0
gensim>=4.3.0  # For word embeddings

# API dependencies (optional)
fastapi>=0.95.0
uvicorn>=0.21.0
